ui : card-component,whats-next,get-docker & translate : whats-next
