+++ public/docs/get-started/get-docker.md
@@ -12,28 +12,31 @@
 >
 > 대기업(직원 250명 이상 또는 연간 매출 1000만 달러 이상)에서 Docker Desktop을 상업적으로 사용하려면 [유료 구독](https://www.docker.com/pricing/?_gl=1*h2v28y*_gcl_au*MjczODgxODI4LjE3Mzg0NzA0NDI.*_ga*MjEyODM1MDY2OC4xNzIwMzEyNzQ5*_ga_XJWPQMJYHQ*MTczOTU2MjU3My42MS4xLjE3Mzk1NjI3NjMuNjAuMC4w)이 필요합니다.
 
-<box-component
+<div class="not-prose">
+<card-component
   imgsrc="/imgs/get-started/get-docker-logo/apple_48.svg"
   title="Docker Desktop for Mac"
   description="macOS 샌드박스 보안 모델을 사용하는 네이티브 애플리케이션으로, Mac에 모든 Docker 도구를 제공합니다."
   href="https://docs.docker.com/desktop/setup/install/mac-install/"
-></box-component>
+></card-component>
 
-
-<box-component
+<card-component
   imgsrc="/imgs/get-started/get-docker-logo/windows_48.svg"
   title="Docker Desktop for Windows"
   description="모든 Docker 도구를 Windows 컴퓨터에 제공하는 네이티브 Windows 애플리케이션입니다."
   href="https://docs.docker.com/desktop/setup/install/windows-install/"
-></box-component>
+></card-component>
 
-
-<box-component
+<card-component
   imgsrc="/imgs/get-started/get-docker-logo/linux_48.svg"
   title="Docker Desktop for Linux"
   description="모든 Docker 도구를 Linux 컴퓨터에 제공하는 네이티브 Linux 애플리케이션입니다."
   href="https://docs.docker.com/desktop/setup/install/linux/"
-></box-component>
+></card-component>
+</div>
+
+
+
 
 > **Note**
 >
