+++ public/docs/get-started/introduction/whats-next.md
@@ -6,10 +6,100 @@
 
 컨테이너, 이미지, 레지스트리, Docker Compose의 핵심 개념을 배워보세요.
 
+<div class="not-prose md:grid-cols-2 lg:grid-cols-3 grid grid-cols-1 gap-4 mb-6 auto-rows-fr">
+<card-component
+title="What is a container?"
+description="첫 번째 컨테이너를 실행하는 방법을 배워보세요."
+href="/#/get-started/docker-concepts/the-basics/what-is-a-container"
+></card-component>
+
+<card-component
+    title="What is an image?"
+    description="이미지 레이어의 기초를 배워보세요."
+    href="/#/get-started/docker-concepts/the-basics/what-is-an-image"
+></card-component>
+
+<card-component
+    title="What is a registry?"
+    description="컨테이너 레지스트리에 대해 알아보고, 레지스트리 간 상호 운용성을 살펴본 후, 실제로 레지스트리와 상호작용해 보세요."
+    href="/#/get-started/docker-concepts/the-basics/what-is-a-registry"
+></card-component>
+
+<card-component
+    title="What is Docker Compose?"
+    description="Docker Compose를 더 잘 이해해 보세요."
+    href="/#/get-started/docker-concepts/the-basics/what-is-docker-compose"
+></card-component>
+</div>
+
 ## Building images
 
 Dockerfiles, 빌드 캐시 및 다단계 빌드를 사용하여 최적화된 컨테이너 이미지를 만들어보세요.
 
+<div class="not-prose md:grid-cols-2 lg:grid-cols-3 grid grid-cols-1 gap-4 mb-6 auto-rows-fr">
+<card-component
+title="Understanding image layers"
+description="컨테이너 이미지 레이어에 대해 배워보세요."
+href="/#/get-started/docker-concepts/building-images/understanding-image-layers"
+></card-component>
+
+<card-component
+    title="Writing a Dockerfile"
+    description="Dockerfile을 사용하여 이미지를 생성하는 방법을 배워보세요."
+    href="/#/get-started/docker-concepts/building-images/writing-a-dockerfile"
+></card-component>
+
+<card-component
+    title="Build, tag and publish an image"
+    description="Docker Hub나 다른 레지스트리에 이미지를 빌드하고 태그한 뒤 게시하는 방법을 알아보세요."
+    href="/#/get-started/docker-concepts/building-images/build-tag-and-publish-an-image"
+></card-component>
+
+<card-component
+    title="Using the build cache"
+    description="빌드 캐시가 무엇인지, 어떤 변경이 캐시를 무효화하는지, 그리고 빌드 캐시를 효과과적으로 활용하는 방법을 배워보세요."
+    href="/#/get-started/docker-concepts/building-images/using-the-build-cache"
+></card-component>
+
+<card-component
+    title="Multi-stage builds"
+    description="멀티 스테이지 빌드와 그 이점에 대해 잘 이해해 보세요."
+    href="/#/get-started/docker-concepts/building-images/multi-stage-builds"
+></card-component>
+</div>
+
 ## Running containers
 
 포트 노출, 기본값 재정의, 데이터 유지, 파일 공유, 멀티 컨테이너 애플리케이션 관리를 위한 필수 기술을 익혀보세요.
+
+<div class="not-prose md:grid-cols-2 lg:grid-cols-3 grid grid-cols-1 gap-4 mb-6 auto-rows-fr">
+<card-component
+    title="Publishing ports"
+    description="Docker에서 포트를 게시하고 노출하는 것의 중요성을 이해해 보세요"
+    href="/#/get-started/docker-concepts/running-containers/publishing-ports"
+></card-component>
+
+<card-component
+    title="Overriding container defaults"
+    description="<code>docker run</code> 명령어를 사용해 컨테이너의 기본 설정을 재정의하는 방법을 알아보세요."
+    href="/#/get-started/docker-concepts/running-containers/overriding-container-defaults"
+></card-component>
+
+<card-component
+    title="Persisting container data"
+    description="Docker에서 데이터 지속성이 왜 중요한지 알아보세요."
+    href="/#/get-started/docker-concepts/running-containers/persisting-container-data"
+></card-component>
+
+<card-component
+    title="Sharing local files with containers"
+    description="Docker에서 사용 가능한 다양한 스토리지 옵션과 사용법을 살펴보세요."
+    href="/#/get-started/docker-concepts/running-containers/sharing-local-files"
+></card-component>
+
+<card-component
+    title="Multi-container applications"
+    description="다중 컨테이너 애플리케이션의 중요성과 단일 컨테이너 애플리케이션과의 차이점을 알아보세요."
+    href="/#/get-started/docker-concepts/running-containers/multi-container-applications"
+></card-component>
+</div>
