+++ src/scripts/components/index.ts
@@ -1,4 +1,4 @@
 import './footer-component';
 import './header-component';
-import './box-component';
+import './card-component';
 import './button-component';
