+++ src/scripts/main.ts
@@ -3,7 +3,7 @@ import '../styles/content_style.css';
 import '../styles/not_found.css';
 import '../styles/style.css';
 import './load_md';
-import './components/box-component';
+import './components/card-component';
 import { initializeMarkdownLoader } from './load_md';
 import { initializeNavFn } from './nav';
 import { initializeTableContents } from './table-contents';
