+++ src/styles/style.css
@@ -1,6 +1,14 @@
 @import 'tailwindcss';
 
+@tailwind base;
+@tailwind components;
+@tailwind utilities;
+
+/* Table 관련 기본 스타일 */
 @layer base {
+  card-component {
+    @apply block w-full h-full;
+  }
   table {
     @apply w-3/4 table-auto;
   }
@@ -12,5 +20,39 @@
 
   thead {
     @apply bg-gray-200;
+  }
+  
+}
+
+/* 카드 컴포넌트 스타일 */
+@layer components {
+  .card {
+    @apply h-full w-full min-w-0 bg-white rounded-md border border-gray-200 px-4 py-3 mb-4 transition duration-200 hover:border-gray-300 hover:shadow;
+  }
+  
+  
+
+  .card-link {
+    @apply flex items-start gap-4 h-full no-underline;
+  }
+
+  .card-icon {
+    @apply flex-shrink-0 mt-1;
+  }
+
+  .card-img {
+    @apply w-6 h-6;
+  }
+
+  .card-content {
+    @apply flex-1;
+  }
+
+  .card-description {
+    @apply text-sm text-gray-500 leading-snug;
+  }
+
+  .card-title {
+    @apply text-base font-semibold text-gray-700;
   }
 }
